requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
gitpython==3.1.43
openai==1.35.0
google-cloud-aiplatform==1.53.0
vertexai==1.52.0
pylint==3.2.3
pyyaml==6.0.1
loguru==0.7.2
typer==0.12.3
schedule==1.2.2
